Move BaseModApiClient into its own apis/base.py module and drop unused get_mod_info stub.
Keeps apis/__init__.py to package constants/re-exports only. Bumped apis (1.2.0->1.3.0) and modrinth_api (1.1.1->1.1.2) versions for the structural removal and import-path change respectively.
